requirements.txt
#iosxr_grpc==1.3
grpcio==1.44.0
protobuf==3.20.1
pyyaml==6.0